Fix libstdc++ build by disabling executables in cross-compilation
This adds the --disable-executables flag to the libstdc++ configure options
to prevent the build from running executable tests, which aren't allowed
when GCC_NO_EXECUTABLES is set during cross-compilation. This resolves
the 'Link tests are not allowed after GCC_NO_EXECUTABLES' error that was
occurring during the Pedigree cross-compiler build.

Fixes the libstdc++ configuration for cross-compilation by adding --disable-executables flag